+++ cypress/integration/core-opensearch-dashboards/opensearch-dashboards/apps/query_enhancements/a_check.spec.js
@@ -9,10 +9,11 @@ import { SECONDARY_ENGINE } from '../../../../../utils/constants';
 
 const miscUtils = new MiscUtils(cy);
 
-describe('Workspace Commands / API Test', () => {
+describe('No Index Pattern Check Test', () => {
   before(() => {
     // Load test data
     cy.setupTestData(
+      SECONDARY_ENGINE.url,
       [
         'cypress/fixtures/query_enhancements/data-logs-1/data_logs_small_time_1.mapping.json',
         'cypress/fixtures/query_enhancements/data-logs-2/data_logs_small_time_2.mapping.json',
